fix: order dto

+++ core/dto/order.go
@@ -33,17 +33,17 @@ type (
 		// 商品数量
 		ItemCount int `json:"itemCount"`
 		// 商品总金额
-		ItemAmount int64 `json:"itemAmount"`
+		ItemAmount int32 `json:"itemAmount"`
 		// 抵扣金额
-		DiscountAmount int64 `json:"discountAmount"`
+		DiscountAmount int32 `json:"discountAmount"`
 		// 优惠金额
-		DeductAmount int64 `json:"deductAmount"`
+		DeductAmount int32 `json:"deductAmount"`
 		// 快递费
-		ExpressFee int64 `json:"expressFee"`
+		ExpressFee int32 `json:"expressFee"`
 		// 包装费
-		PackageFee int64 `json:"packageFee"`
+		PackageFee int32 `json:"packageFee"`
 		// 最终金额
-		FinalAmount int64 `json:"finalAmount"`
+		FinalAmount int32 `json:"finalAmount"`
 		// 是否支付
 		IsPaid int32 `json:"isPaid"`
 		// 状态
